+++ client/flash.h
@@ -11,6 +11,7 @@
 
 #include <stdint.h>
 #include "elf.h"
+#include "uart.h"
 
 typedef struct {
 	void *data;
@@ -26,10 +27,13 @@ typedef struct {
 } flash_file_t;
 
 int flash_load(flash_file_t *ctx, const char *name, int can_write_bl);
-int flash_start_flashing(int enable_bl_writes,char *serial_port_name);
+int flash_start_flashing(int enable_bl_writes, char *serial_port_name);
 int flash_write(flash_file_t *ctx);
 void flash_free(flash_file_t *ctx);
 int flash_stop_flashing(void);
+void CloseProxmark(const char *serial_port_name);
+int OpenProxmark(size_t i, const char *serial_port_name);
 
+extern serial_port sp;
 #endif
